Title: 실수 받아 그대로 출력, Time: 143ms, Memory: 11MB, Status: Passed - Codetree

+++ 250723/실수 받아 그대로 출력/enter-real-value-and-print.java
@@ -3,7 +3,7 @@ public class Main {
     public static void main(String[] args) {
         // Please write your code here.
         Scanner sc = new Scanner(System.in);
-        double n = sc.nextInt();
+        double n = sc.nextDouble();
         System.out.printf("%.2f", n);
     }
 }
